requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Supabase
supabase>=2.3.0
sqlalchemy==2.0.23
alembic==1.13.0

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Background Jobs
celery==5.3.4

# HTTP Client (compatible with supabase 2.27.2 and gotrue 2.9.1)
httpx>=0.26.0,<0.28.0

# WebSockets (required by supabase realtime)
websockets>=13.0,<17.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
pillow-heif==0.13.0  # HEIC/HEIF support

# AI/ML
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
faiss-cpu==1.7.4

# Color Analysis
colorthief==0.2.1
extcolors==1.0.0
colormath==3.0.0
scikit-image==0.22.0

# NLP & Templates
jinja2==3.1.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Utils
email-validator==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
